Move Figure.shift_origin to a standalone Python file
