+++ pygmt/src/__init__.py
@@ -43,6 +43,7 @@ from pygmt.src.plot3d import plot3d
 from pygmt.src.project import project
 from pygmt.src.rose import rose
 from pygmt.src.select import select
+from pygmt.src.shift_origin import shift_origin
 from pygmt.src.solar import solar
 from pygmt.src.sph2grd import sph2grd
 from pygmt.src.sphdistance import sphdistance
